+++ tests/ipp-tests.sh
@@ -2,7 +2,7 @@
 #
 # IPP Everywhere Printer Self-Certification Manual 1.1: Section 6: IPP Tests.
 #
-# Copyright 2014-2020 by the IEEE-ISTO Printer Working Group.
+# Copyright 2014-2022 by the IEEE-ISTO Printer Working Group.
 #
 # Licensed under Apache License v2.0.  See the file "LICENSE" for more
 # information.
@@ -46,7 +46,7 @@ done
 
 PLIST="${TARGET} IPP Results.plist"
 
-"${IPPFIND}" --literal-name "${TARGET}" -x "${IPPTOOL}" -P "${PLIST}" -I '{}' ipp-tests.test \;
+"${IPPFIND}" --literal-name "${TARGET}" -x "${IPPTOOL}" -P "${PLIST}" -I -T 120 '{}' ipp-tests.test \;
 
 # confirm that the PLIST is well formed, if plutil is available (e.g. running on Darwin / OS X / macOS)
 test `which plutil` && plutil -lint -s "${PLIST}"